+++ R/ip_interface.R
@@ -30,6 +30,9 @@ methods::setOldClass(c("ip_interface", "vctrs_vctr"))
 #' For other purposes, you can extract the address and network components using
 #' [as_ip_address()] and [as_ip_network()].
 #'
+#' When comparing and sorting `ip_interface` vectors, the network is compared
+#' before the host address.
+#'
 #' @param ... Included for S3 generic consistency
 #' @param x
 #' * For `ip_interface()`: A character vector of IP interfaces, in CIDR notation
@@ -137,6 +140,17 @@ as.character.ip_interface <- function(x, ...) wrap_print_network(x)
 #' @rdname ip_interface
 #' @export
 format.ip_interface <- function(x, ...) as.character(x)
+
+
+# Comparison ------------------------------------------------------------
+
+#' @export
+vec_proxy_compare.ip_interface <- function(x, ...) {
+  compare_network <- vec_proxy_compare(as_ip_network(x))
+  compare_address <- vec_proxy_compare(as_ip_address(x))
+
+  cbind(compare_network,  compare_address)
+}
 
 
 # Other ------------------------------------------------------------